refactor(providers): split repository providers out of app_providers.dart
Second sub-batch (9b of 3) of #4506's final wave. Refs #4339 Wave 2.

Moves the 14-entity repository cluster to lib/providers/repository_providers.dart:

- badgeRepositoryProvider (legacy Provider<X>)
- cachedFollowingList + followRepository (follow cluster)
- profileRepository, categoriesRepository, hashtagRepository
- curatedListRepository + CuratedListsState (notifier) + curationRepository
- peopleListsRepository (+ private _peopleListsBoxName const)
- bookmarkService, muteService (NIP-51 services with repository shape)
- dmRepository (NIP-17 DM persistence)
- commentsRepository (NIP-22 comments)

Migration pattern (mirrors batches 5-9a):

  // app_providers.dart
  import 'package:openvine/providers/repository_providers.dart';
  // ...
  export 'repository_providers.dart';

- Back-import keeps createBlockedAuthorFilter + pendingActionServiceProvider
  (both stay in app_providers.dart until 9c) resolvable from the new file.
- Export keeps existing consumers' imports unchanged.

Scope cleanup:

- dart fix --apply removed 23 newly-unused imports from app_providers.dart.
- Three dead-code marker comments deleted: VideoManagerService removed,
  Legacy ExploreVideoManager removed, "In app_providers.dart" stub.
- Four workspace package dependencies surfaced as direct deps in pubspec.yaml
  (comments_repository, curated_list_repository, profile_repository, models),
  added bare in the workspace-deps section to match existing convention.

Cumulative reduction: app_providers.dart 978 -> 455 LoC (84% reduction from
the original 2,918). One sub-batch remains (9c: social + foundation) to
fully exhaust #4506.

Test plan:
- dart run build_runner build --delete-conflicting-outputs (clean)
- flutter analyze lib test integration_test (no issues)
- flutter test test/screens/feed/pooled_video_feed_item_repo_swap_test.dart
  test/widgets/video_feed_item/feed_videos_repo_swap_test.dart
  test/providers/individual_video_providers_test.dart
  test/providers/dm_repository_provider_test.dart
  test/screens/comments/comments_header_test.dart (16 passed)
